Change page 08 in level 1 to say 'choose'
I think Alex was going to use the obtain tactic and then changed to
using 'choose' instead.  So the page where 'choose' is introduced is
called 'obtain' and mentions how you will use 'obtain' all over the
place.  But 'obtain' isn't a tactic that is available at this point.

Change the wording to (I hope) preserve the intent but make it less
confusing.  Also change the page name to choose so that's consistent.